+++ packages/retail-ui/components/Fias/Fias.tsx
@@ -14,7 +14,7 @@ import isEqual from 'lodash.isequal';
 import { Logger } from './logger/Logger';
 import { cx } from '../../lib/theming/Emotion';
 import jsStyles from './Fias.styles';
-import { ThemeConsumer } from '../internal/ThemeContext';
+import { ThemeConsumer } from '../ThemeConsumer';
 import { ITheme } from '../../lib/theming/Theme';
 
 export interface FiasProps {
