+++ packages/retail-ui/components/Kebab/Kebab.tsx
@@ -12,7 +12,7 @@ import styles from './Kebab.module.less';
 import { PopupPosition } from '../Popup';
 import { cx } from '../../lib/theming/Emotion';
 import jsStyles from './Kebab.styles';
-import { ThemeConsumer } from '../internal/ThemeContext';
+import { ThemeConsumer } from '../ThemeConsumer';
 import { ITheme } from '../../lib/theming/Theme';
 
 export interface KebabProps {
