+++ packages/retail-ui/components/Menu/Menu.tsx
@@ -8,7 +8,7 @@ import styles from './Menu.module.less';
 import { Nullable } from '../../typings/utility-types';
 import { cx } from '../../lib/theming/Emotion';
 import jsStyles from './Menu.styles';
-import { ThemeConsumer } from '../internal/ThemeContext';
+import { ThemeConsumer } from '../ThemeConsumer';
 import { ITheme } from '../../lib/theming/Theme';
 
 interface MenuProps {
