+++ packages/retail-ui/components/MenuSeparator/MenuSeparator.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import styles from './MenuSeparator.module.less';
 import { cx } from '../../lib/theming/Emotion';
 import jsStyles from './MenuSeparator.styles';
-import { ThemeConsumer } from '../internal/ThemeContext';
+import { ThemeConsumer } from '../ThemeConsumer';
 import { ITheme } from '../../lib/theming/Theme';
 
 /**
