+++ packages/retail-ui/components/Modal/Modal.tsx
@@ -19,7 +19,7 @@ import { isIE } from '../ensureOldIEClassName';
 import styles from './Modal.module.less';
 import { cx } from '../../lib/theming/Emotion';
 import jsStyles from './Modal.styles';
-import { ThemeConsumer } from '../internal/ThemeContext';
+import { ThemeConsumer } from '../ThemeConsumer';
 import { ITheme } from '../../lib/theming/Theme';
 
 let mountedModalsCount = 0;
