+++ packages/retail-ui/components/Modal/ModalClose.tsx
@@ -3,7 +3,7 @@ import { CloseProps } from './ModalContext';
 import styles from './Modal.module.less';
 import { cx } from '../../lib/theming/Emotion';
 import jsStyles from './Modal.styles';
-import { ThemeConsumer } from '../internal/ThemeContext';
+import { ThemeConsumer } from '../ThemeConsumer';
 import { ITheme } from '../../lib/theming/Theme';
 
 export default class Close extends React.Component<CloseProps> {
