+++ packages/retail-ui/components/Switcher/Switcher.tsx
@@ -7,7 +7,7 @@ import styles from './Switcher.module.less';
 import { Nullable } from '../../typings/utility-types';
 import { cx } from '../../lib/theming/Emotion';
 import jsStyles from './Switcher.styles';
-import { ThemeConsumer } from '../internal/ThemeContext';
+import { ThemeConsumer } from '../ThemeConsumer';
 import { ITheme } from '../../lib/theming/Theme';
 
 export type SwitcherSize = ButtonSize;
