+++ packages/retail-ui/components/Tabs/Tab.tsx
@@ -9,7 +9,7 @@ import styles from './Tab.module.less';
 import { TabsContextType, TabsContext } from './TabsContext';
 import { cx } from '../../lib/theming/Emotion';
 import jsStyles from './Tab.styles';
-import { ThemeConsumer } from '../internal/ThemeContext';
+import { ThemeConsumer } from '../ThemeConsumer';
 import { ITheme } from '../../lib/theming/Theme';
 
 export interface TabIndicators {
