+++ packages/retail-ui/components/ThemeConsumer/ThemeConsumer.tsx
@@ -1,1 +1,1 @@
-export { ThemeConsumer, ThemeConsumer as default } from '../internal/ThemeContext';
+export { ThemeConsumer, ThemeConsumer as default } from '../../lib/theming/ThemeContext';
