+++ packages/retail-ui/components/Tooltip/Tooltip.tsx
@@ -10,7 +10,7 @@ import isEqual from 'lodash.isequal';
 import { containsTargetOrRenderContainer } from '../../lib/listenFocusOutside';
 import jsStyles from './Tooltip.styles';
 import { cx } from '../../lib/theming/Emotion';
-import { ThemeConsumer } from '../internal/ThemeContext';
+import { ThemeConsumer } from '../ThemeConsumer';
 import { ITheme } from '../../lib/theming/Theme';
 const POPUP_MARGIN = 15;
 const POPUP_PIN_OFFSET = 17;
