+++ packages/retail-ui/components/TopBar/TopBar.tsx
@@ -14,7 +14,7 @@ import Start from './TopBarStart';
 import Logout from './TopBarLogout';
 import { cx } from '../../lib/theming/Emotion';
 import jsStyles from './TopBar.styles';
-import { ThemeConsumer } from '../internal/ThemeContext';
+import { ThemeConsumer } from '../ThemeConsumer';
 import { ITheme } from '../../lib/theming/Theme';
 
 export interface TopBarProps {
