+++ packages/retail-ui/components/internal/MaskedInput/MaskedInput.tsx
@@ -3,7 +3,7 @@ import ReactInputMask, { InputState, MaskOptions } from 'react-input-mask';
 import styles from './MaskedInput.module.less';
 import jsStyles from './MaskedInput.styles';
 import { cx } from '../../../lib/theming/Emotion';
-import { ThemeConsumer } from '../ThemeContext';
+import { ThemeConsumer } from '../../ThemeConsumer';
 import { ITheme } from '../../../lib/theming/Theme';
 
 export interface MaskedInputProps extends React.InputHTMLAttributes<HTMLInputElement> {
