--- packages/retail-ui/components/internal/ThemeContext.ts
@@ -1,4 +1,0 @@
-import ThemeContext from '../../lib/theming/ThemeContext';
-
-export const ThemeConsumer = ThemeContext.Consumer;
-export const ThemeProvider = ThemeContext.Provider;
