+++ packages/retail-ui/lib/theming/ThemeContext.ts
@@ -8,4 +8,6 @@ import ThemeFactory from './ThemeFactory';
 // to get it compatible with the useContext hook
 export const ThemeContext = createReactContext(ThemeFactory.getDefaultTheme()) as React.Context<ITheme>;
 
-export default ThemeContext
+export default ThemeContext;
+export const ThemeProvider = ThemeContext.Provider;
+export const ThemeConsumer = ThemeContext.Consumer;
